+++ eslint-suppressions.json
@@ -1536,7 +1536,7 @@
   },
   "public/app/features/dashboard-scene/edit-pane/shared.ts": {
     "no-barrel-files/no-barrel-files": {
-      "count": 3
+      "count": 1
     }
   },
   "public/app/features/dashboard-scene/inspect/HelpWizard/HelpWizard.tsx": {
