+++ public/app/features/dashboard-scene/edit-pane/dashboardEditActions/addVariable.ts
@@ -1,0 +1,17 @@
+import { addElement } from './addElement';
+import { type AddVariableActionHelperProps } from './types';
+
+export function addVariable({ source, addedObject }: AddVariableActionHelperProps) {
+  const varsBeforeAddition = [...(source.state.variables ?? [])];
+
+  addElement({
+    source,
+    addedObject,
+    perform() {
+      source.setState({ variables: [...varsBeforeAddition, addedObject] });
+    },
+    undo() {
+      source.setState({ variables: [...varsBeforeAddition] });
+    },
+  });
+}
