+++ public/app/features/dashboard-scene/edit-pane/dashboardEditActions/changeVariableLabel.ts
@@ -1,0 +1,10 @@
+/* eslint-disable @grafana/i18n/no-translation-top-level */
+import { t } from '@grafana/i18n';
+import { type SceneVariable } from '@grafana/scenes';
+
+import { makeEditAction } from './makeEditAction';
+
+export const changeVariableLabel = makeEditAction<SceneVariable, 'label'>({
+  description: t('dashboard.variable.label.action', 'Change variable label'),
+  prop: 'label',
+});
